쿠폰 관리 기능 별 컴포넌트 세분화
