+++ src/refactoring/components/features/coupon/CouponItem.tsx
@@ -1,0 +1,17 @@
+import { Coupon } from "../../../../types";
+
+type CouponItemProps = {
+  coupon: Coupon;
+};
+
+export const CouponItem = ({ coupon, ...props }: CouponItemProps) => {
+  return (
+    <div className="bg-gray-100 p-2 rounded" {...props}>
+      {coupon.name} ({coupon.code}):
+      {coupon.discountType === "amount"
+        ? `${coupon.discountValue}원`
+        : `${coupon.discountValue}%`}{" "}
+      할인
+    </div>
+  );
+};
