+++ src/refactoring/models/coupon.ts
@@ -1,0 +1,8 @@
+import { Coupon } from "../../types";
+
+export const createNewCoupon = (): Coupon => ({
+  name: "",
+  code: "",
+  discountType: "percentage",
+  discountValue: 0,
+});
